ADD: created ListItem component
