+++ src/components/List/List.js
@@ -1,105 +1,13 @@
 import React, { useState } from 'react';
-import { doc, updateDoc } from 'firebase/firestore';
 import { useParams, Link } from 'react-router-dom';
-import { db } from '../../lib/firebase';
-import { calculateEstimate } from '@the-collab-lab/shopping-list-utils';
 import useSnapshot from '../../hooks/useSnapshot';
 import Navigation from '../Navigation/Navigation';
-import { DAY_IN_MILLISEC } from '../../lib/util';
-
-export const daysSincePurchase = (datePurchaseInMilli, dateCreatedInMilli) => {
-  const workingTimestamp = datePurchaseInMilli || dateCreatedInMilli;
-
-  const differenceInMilli = Date.now() - workingTimestamp;
-  return differenceInMilli / DAY_IN_MILLISEC;
-};
-
-function getPurchaseDates(estimateInDays) {
-  const purchaseDate = new Date();
-  const estimatedNextPurchaseDate = new Date(
-    purchaseDate.getTime() + estimateInDays * DAY_IN_MILLISEC,
-  );
-
-  return [purchaseDate, estimatedNextPurchaseDate];
-}
-
-function getPurchaseData(item) {
-  const totalPurchases = item.totalPurchases + 1;
-  const purchaseFreq = calculateEstimate(
-    item.purchaseFreq,
-    daysSincePurchase(item.purchaseDate?.toMillis(), item.createdAt.toMillis()),
-    totalPurchases,
-  );
-
-  return [totalPurchases, purchaseFreq];
-}
+import ListItem from '../ListItem/ListItem';
 
 export default function List() {
   const { token } = useParams();
   const { docs } = useSnapshot(token);
   const [userSearch, setUserSearch] = useState('');
-
-  async function undoPurchase(id) {
-    await updateDoc(doc(db, token, id), {
-      purchaseDate: null,
-      estimatedNextPurchaseDate: null,
-    });
-  }
-
-  async function updatePurchase(item) {
-    const [totalPurchases, purchaseFreq] = getPurchaseData(item);
-    const [purchaseDate, estimatedNextPurchaseDate] =
-      getPurchaseDates(purchaseFreq);
-
-    await updateDoc(doc(db, token, item.id), {
-      purchaseDate,
-      totalPurchases,
-      purchaseFreq,
-      estimatedNextPurchaseDate,
-    });
-  }
-
-  function checkboxChange(item) {
-    if (item.checked) {
-      undoPurchase(item.id);
-    } else {
-      updatePurchase(item);
-    }
-  }
-
-  function isInactive(totalPurchases, purchaseDate, estimatedNextPurchaseDate) {
-    if (totalPurchases < 2) return true;
-    const date = new Date(purchaseDate);
-    const timeElapsed = date.getTime() - Date.now();
-    const estiDate = new Date(estimatedNextPurchaseDate);
-    const comparison = timeElapsed * 2 > estiDate.getTime();
-
-    return comparison;
-  }
-
-  function styleCheckbox(item) {
-    const inactive = isInactive(
-      item.totalPurchases,
-      item.purchaseDate,
-      item.estimatedNextPurchaseDate,
-    );
-    if (inactive) return 'black';
-    if (item.purchaseFreq <= 7) return 'green';
-    if (item.purchaseFreq <= 14) return 'orange';
-    return 'red';
-  }
-
-  function accessibilityLabel(item) {
-    const inactive = isInactive(
-      item.totalPurchases,
-      item.purchaseDate,
-      item.estimatedNextPurchaseDate,
-    );
-    if (inactive) return 'not purchase';
-    if (item.purchaseFreq <= 7) return 'purchase in less than 7 days';
-    if (item.purchaseFreq <= 14) return 'purchase in less than 14 days';
-    return 'purchase in 30 days';
-  }
 
   return (
     <>
@@ -154,26 +62,7 @@ export default function List() {
                 return 0;
               })
               .map((doc) => {
-                return (
-                  <li key={doc.id}>
-                    <input
-                      key={`checkbox-${doc.id}`}
-                      type="checkbox"
-                      id={doc.id}
-                      checked={doc.checked}
-                      onChange={() => checkboxChange(doc)}
-                    />
-                    <span
-                      key={`item-${doc.id}`}
-                      aria-label={accessibilityLabel(doc)}
-                      style={{
-                        color: styleCheckbox(doc),
-                      }}
-                    >
-                      {doc.item}
-                    </span>
-                  </li>
-                );
+                return <ListItem commodity={doc} token={token} />;
               })}
           </ul>
         </>
